requirements.txt
absl-py==2.1.0
altair==5.3.0
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/altgraph-0.17.2-py2.py3-none-any.whl
astunparse==1.6.3
attrs==23.2.0
blinker==1.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
eli5==0.13.0
flatbuffers==24.3.25
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/future-0.18.2-py3-none-any.whl
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.63.0
h5py==3.11.0
idna==3.6
imbalanced-learn==0.12.2
imblearn==0.0
importlib_metadata==7.1.0
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==3.3.3
libclang==18.1.1
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mysql-connector-python==8.3.0
names==0.3.0
namex==0.0.8
noaa-sdk==0.1.21
numpy==1.26.2
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.1.3
pillow==10.3.0
protobuf==4.25.3
pyarrow==16.0.0
pydeck==0.9.0
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.2
scipy==1.13.0
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/six-1.15.0-py2.py3-none-any.whl
smmap==5.0.1
streamlit==1.34.0
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
threadpoolctl==3.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.16.0
zipp==3.18.1
